+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## next
+
+- Support for simple types
+- Support for tuples
+- Group error type added
+
 ## v0.1.0 (2024-01-09)
 
 - Initial release